+++ CHANGELOG.rst
@@ -6,6 +6,7 @@
 =======
 
  * added option to identify targets as Precious in SlurmEnvironment
+ * added a SlurmEnvironment time option to output linux system time stats per Command
 
 0.6
 ===